+++ src/bandcamp/content.js
@@ -1,5 +1,5 @@
 import { log } from '../utils/console';
-import { injectCSSFile } from '../utils/html';
+import { injectCSSFile, injectJSFile } from '../utils/html';
 import { PageTypeDetector } from './app/page-type.js';
 import { setupPageAlbum } from './pages/page-album.js';
 import { setupPageMusic } from './pages/page-music.js';
@@ -9,13 +9,14 @@ import './css/b2d.css';
 function main() {
   log('Running bandcamp content main module logic (bandcamp/modules/main.js)');
   injectCSSFile(chrome.runtime.getURL('bandcamp.content.css'));
+  injectJSFile(chrome.runtime.getURL('bandcamp.inject.js'));
 
   const pageType = new PageTypeDetector().detect();
 
   if (pageType.isMusic()) {
-    setupPageMusic();
+    setupPageMusic(pageType);
   } else if (pageType.isAlbum()) {
-    setupPageAlbum();
+    setupPageAlbum(pageType);
   }
 }