+++ src/popup/modules/main.ts
@@ -3,6 +3,7 @@ import { click } from '../../utils/html';
 import { ReleasesList } from '../components/releases-list';
 import { setupReleaseCardTab } from '../tabs/release-card_tab';
 import { setupReleasesTab } from '../tabs/releases_tab';
+import { Music } from 'src/app/music';
 
 export function showReleaseCardTab(release: Release) {
   const btnReleaseCardTab = getReleaseCardTabElement();
@@ -31,15 +32,14 @@ export function setupReleasesTabElement(): void {
 }
 
 export function showReleasesTabContent(
-  releasesData: [],
-  bgImageSrc: string | undefined,
+  music: Music,
   searchValue: string | undefined
 ): void {
   const btnReleasesTab = getReleasesTabElement();
   if (!btnReleasesTab) return;
 
   click(btnReleasesTab);
-  setupReleasesTab(releasesData, bgImageSrc, searchValue);
+  setupReleasesTab(music, searchValue);
 }
 
 export function getReleaseCardTabElement(): HTMLElement | null {